.github/scripts/requirements.txt
GitPython==3.1.27
loguru==0.6.0
ruamel.yaml==0.17.21
ruamel.yaml.string==0.1.0
typer==0.6.1
